backend/requirements.txt
Django==3.2.3
django-filter==23.3
djangorestframework==3.12.4
djoser==2.2.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
Pillow==9.0.0
PyYAML==6.0
gunicorn==20.1.0
flake8==6.0.0
flake8-isort==6.0.0
python-dotenv==0.19.0
